+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vedtak/VedtakSteg.kt
@@ -14,6 +14,7 @@ import no.nav.familie.ks.sak.kjerne.behandling.steg.BehandlingStegStatus
 import no.nav.familie.ks.sak.kjerne.behandling.steg.IBehandlingSteg
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene.VedtaksperiodeService
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.validerPerioderInneholderBegrunnelser
+import no.nav.familie.ks.sak.kjerne.brev.GenererBrevService
 import no.nav.familie.ks.sak.kjerne.logg.LoggService
 import no.nav.familie.ks.sak.kjerne.totrinnskontroll.TotrinnskontrollService
 import no.nav.familie.prosessering.internal.TaskService
@@ -22,6 +23,7 @@ import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDate
+import java.time.LocalDateTime
 
 @Service
 class VedtakSteg(
@@ -31,7 +33,8 @@ class VedtakSteg(
     private val loggService: LoggService,
     private val oppgaveService: OppgaveService,
     private val vedtakService: VedtakService,
-    private val vedtaksperiodeService: VedtaksperiodeService
+    private val vedtaksperiodeService: VedtaksperiodeService,
+    private val genererBrevService: GenererBrevService
 ) : IBehandlingSteg {
     override fun getBehandlingssteg(): BehandlingSteg = BehandlingSteg.VEDTAK
 
@@ -54,6 +57,16 @@ class VedtakSteg(
         taskService.save(godkjenneVedtakTask)
 
         opprettFerdigstillOppgaveTasker(behandling)
+
+        val vedtak = vedtakService.hentAktivVedtakForBehandling(behandlingId)
+
+        vedtak.vedtaksdato = LocalDateTime.now()
+        if (behandling.skalSendeVedtaksbrev()) {
+            val brev = genererBrevService.genererBrevForBehandling(behandling.id)
+            vedtak.stønadBrevPdf = brev
+        }
+
+        vedtakService.oppdaterVedtak(vedtak)
 
         behandlingService.oppdaterStatusPåBehandling(behandlingId, BehandlingStatus.FATTER_VEDTAK)
     }
