+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/brev/BrevService.kt
@@ -7,9 +7,6 @@ import no.nav.familie.kontrakter.felles.dokarkiv.v2.Filtype
 import no.nav.familie.kontrakter.felles.dokarkiv.v2.Førsteside
 import no.nav.familie.ks.sak.api.dto.DistribuerBrevDto
 import no.nav.familie.ks.sak.api.dto.ManueltBrevDto
-import no.nav.familie.ks.sak.api.dto.tilBrev
-import no.nav.familie.ks.sak.common.exception.Feil
-import no.nav.familie.ks.sak.common.exception.FunksjonellFeil
 import no.nav.familie.ks.sak.config.BehandlerRolle
 import no.nav.familie.ks.sak.integrasjon.distribuering.DistribuerBrevTask
 import no.nav.familie.ks.sak.integrasjon.distribuering.DistribuerDødsfallBrevPåFagsakTask
@@ -20,36 +17,16 @@ import no.nav.familie.ks.sak.integrasjon.journalføring.domene.DbJournalpost
 import no.nav.familie.ks.sak.integrasjon.journalføring.domene.DbJournalpostType
 import no.nav.familie.ks.sak.integrasjon.journalføring.domene.JournalføringRepository
 import no.nav.familie.ks.sak.integrasjon.logger
-import no.nav.familie.ks.sak.integrasjon.sanity.SanityService
-import no.nav.familie.ks.sak.integrasjon.sanity.domene.SanityBegrunnelse
 import no.nav.familie.ks.sak.kjerne.arbeidsfordeling.ArbeidsfordelingService
 import no.nav.familie.ks.sak.kjerne.behandling.SettBehandlingPåVentService
 import no.nav.familie.ks.sak.kjerne.behandling.domene.Behandling
 import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingRepository
-import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingÅrsak
-import no.nav.familie.ks.sak.kjerne.behandling.steg.BehandlingSteg
-import no.nav.familie.ks.sak.kjerne.behandling.steg.simulering.SimuleringService
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.VedtakService
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.domene.Vedtak
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene.UtvidetVedtaksperiodeMedBegrunnelser
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene.VedtaksperiodeService
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.VilkårsvurderingService
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.AnnenVurderingType
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Resultat
-import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.tilSanityBegrunnelse
-import no.nav.familie.ks.sak.kjerne.brev.domene.FellesdataForVedtaksbrev
-import no.nav.familie.ks.sak.kjerne.brev.domene.VedtaksbrevDto
 import no.nav.familie.ks.sak.kjerne.brev.domene.maler.Brevmal
-import no.nav.familie.ks.sak.kjerne.brev.domene.maler.Førstegangsvedtak
-import no.nav.familie.ks.sak.kjerne.brev.domene.maler.Hjemmeltekst
 import no.nav.familie.ks.sak.kjerne.logg.LoggService
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.PersonopplysningGrunnlagService
-import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.Målform
-import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.PersonopplysningGrunnlag
-import no.nav.familie.ks.sak.kjerne.totrinnskontroll.TotrinnskontrollService
-import no.nav.familie.ks.sak.sikkerhet.SikkerhetContext
 import no.nav.familie.prosessering.internal.TaskService
-import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDate
@@ -57,7 +34,6 @@ import java.util.Properties
 
 @Service
 class BrevService(
-    private val brevKlient: BrevKlient,
     private val integrasjonClient: IntegrasjonClient,
     private val loggService: LoggService,
     private val taskService: TaskService,
@@ -68,18 +44,12 @@ class BrevService(
     private val behandlingRepository: BehandlingRepository,
     private val journalføringRepository: JournalføringRepository,
     private val settBehandlingPåVentService: SettBehandlingPåVentService,
-    private val vedtaksperiodeService: VedtaksperiodeService,
-    private val brevPeriodeService: BrevPeriodeService,
-    private val sanityService: SanityService,
-    private val totrinnskontrollService: TotrinnskontrollService,
-    private val vedtakService: VedtakService,
-    private val simuleringService: SimuleringService
-
+    private val genererBrevService: GenererBrevService
 ) {
 
     fun hentForhåndsvisningAvBrev(behandlingId: Long, manueltBrevDto: ManueltBrevDto): ByteArray {
         val manueltBrevDtoMedMottakerData = utvidManueltBrevDtoMedEnhetOgMottaker(behandlingId, manueltBrevDto)
-        return genererManueltBrev(manueltBrevDtoMedMottakerData, true)
+        return genererBrevService.genererManueltBrev(manueltBrevDtoMedMottakerData, true)
     }
 
     fun genererOgSendBrev(behandlingId: Long, manueltBrevDto: ManueltBrevDto) {
@@ -104,62 +74,11 @@ class BrevService(
         )
     }
 
-    private fun genererManueltBrev(
-        manueltBrevRequest: ManueltBrevDto,
-        erForhåndsvisning: Boolean = false
-    ): ByteArray {
-        try {
-            val brev = manueltBrevRequest.tilBrev()
-            return brevKlient.genererBrev(
-                målform = manueltBrevRequest.mottakerMålform.tilSanityFormat(),
-                brev = brev
-            )
-        } catch (it: Exception) {
-            if (it is Feil || it is FunksjonellFeil) {
-                throw it
-            } else {
-                throw Feil(
-                    message = "Klarte ikke generere brev for ${manueltBrevRequest.brevmal}. ${it.message}",
-                    frontendFeilmelding = "${if (erForhåndsvisning) "Det har skjedd en feil" else "Det har skjedd en feil, og brevet er ikke sendt"}. Prøv igjen, og ta kontakt med brukerstøtte hvis problemet vedvarer.",
-                    httpStatus = HttpStatus.INTERNAL_SERVER_ERROR,
-                    throwable = it
-                )
-            }
-        }
-    }
-
-    fun genererBrevForBehandling(behandlingId: Long): ByteArray {
-        val vedtak = vedtakService.hentAktivVedtakForBehandling(behandlingId)
-        try {
-            if (vedtak.behandling.steg > BehandlingSteg.BESLUTTE_VEDTAK) {
-                throw FunksjonellFeil("Ikke tillatt å generere brev etter at behandlingen er sendt fra beslutter")
-            }
-
-            val målform = personopplysningGrunnlagService.hentSøkersMålform(vedtak.behandling.id)
-            val vedtaksbrev =
-                when (vedtak.behandling.opprettetÅrsak) {
-                    BehandlingÅrsak.DØDSFALL -> TODO() // brevService.hentDødsfallbrevData(vedtak)
-                    BehandlingÅrsak.KORREKSJON_VEDTAKSBREV -> TODO() // brevService.hentKorreksjonbrevData(vedtak)
-                    else -> hentVedtaksbrevData(vedtak)
-                }
-            return brevKlient.genererBrev(målform.tilSanityFormat(), vedtaksbrev)
-        } catch (feil: Exception) {
-            if (feil is FunksjonellFeil) throw feil
-
-            throw Feil(
-                message = "Klarte ikke generere vedtaksbrev på behandling ${vedtak.behandling}: ${feil.message}",
-                frontendFeilmelding = "Det har skjedd en feil, og brevet er ikke sendt. Prøv igjen, og ta kontakt med brukerstøtte hvis problemet vedvarer.",
-                httpStatus = HttpStatus.INTERNAL_SERVER_ERROR,
-                throwable = feil
-            )
-        }
-    }
-
     @Transactional
     fun sendBrev(behandlingId: Long, manueltBrevDto: ManueltBrevDto) {
         val behandling = behandlingRepository.hentBehandling(behandlingId)
 
-        val generertBrev = genererManueltBrev(manueltBrevDto, false)
+        val generertBrev = genererBrevService.genererManueltBrev(manueltBrevDto, false)
 
         val førsteside = if (manueltBrevDto.brevmal.skalGenerereForside()) {
             Førsteside(
@@ -314,95 +233,4 @@ class BrevService(
         vilkårsvurdering.personResultater.single { it.erSøkersResultater() }
             .leggTilBlankAnnenVurdering(AnnenVurderingType.OPPLYSNINGSPLIKT)
     }
-
-    fun hentVedtaksbrevData(vedtak: Vedtak): VedtaksbrevDto {
-        val brevtype = hentVedtaksbrevmal(vedtak.behandling)
-        val fellestdataForVedtaksbrev = lagDataForVedtaksbrev(vedtak)
-
-        return when (brevtype) {
-            Brevmal.VEDTAK_FØRSTEGANGSVEDTAK -> Førstegangsvedtak(
-                fellesdataForVedtaksbrev = fellestdataForVedtaksbrev,
-                etterbetaling = simuleringService.hentEtterbetaling(vedtak.behandling.id)
-            )
-
-            else -> throw Feil("Forsøker å hente vedtaksbrevdata for brevmal ${brevtype.visningsTekst}")
-        }
-    }
-
-    fun lagDataForVedtaksbrev(vedtak: Vedtak): FellesdataForVedtaksbrev {
-        val utvidetVedtaksperioderMedBegrunnelser =
-            vedtaksperiodeService.hentUtvidetVedtaksperioderMedBegrunnelser(vedtak).filter {
-                !(it.begrunnelser.isEmpty() && it.fritekster.isEmpty() && it.eøsBegrunnelser.isEmpty())
-            }
-
-        if (utvidetVedtaksperioderMedBegrunnelser.isEmpty()) {
-            throw FunksjonellFeil(
-                "Vedtaket mangler begrunnelser. Du må legge til begrunnelser for å generere vedtaksbrevet."
-            )
-        }
-
-        val personopplysningsgrunnlagOgSignaturData = hentGrunnlagOgSignaturData(vedtak)
-
-        val brevPeriodeDtoer = brevPeriodeService
-            .hentBrevPeriodeDtoer(utvidetVedtaksperioderMedBegrunnelser, vedtak.behandling.id)
-
-        val hjemler = hentHjemler(
-            behandlingId = vedtak.behandling.id,
-            utvidetVedtaksperioderMedBegrunnelser = utvidetVedtaksperioderMedBegrunnelser,
-            målform = personopplysningsgrunnlagOgSignaturData.grunnlag.søker.målform,
-            sanityBegrunnelser = sanityService.hentSanityBegrunnelser()
-        )
-
-        return FellesdataForVedtaksbrev(
-            enhet = personopplysningsgrunnlagOgSignaturData.enhet,
-            saksbehandler = personopplysningsgrunnlagOgSignaturData.saksbehandler,
-            beslutter = personopplysningsgrunnlagOgSignaturData.beslutter,
-            hjemmeltekst = Hjemmeltekst(hjemler),
-            søkerNavn = personopplysningsgrunnlagOgSignaturData.grunnlag.søker.navn,
-            søkerFødselsnummer = personopplysningsgrunnlagOgSignaturData.grunnlag.søker.aktør.aktivFødselsnummer(),
-            perioder = brevPeriodeDtoer,
-            gjelder = personopplysningsgrunnlagOgSignaturData.grunnlag.søker.navn
-        )
-    }
-
-    private fun hentGrunnlagOgSignaturData(vedtak: Vedtak): GrunnlagOgSignaturData {
-        val personopplysningGrunnlag =
-            personopplysningGrunnlagService.hentAktivPersonopplysningGrunnlagThrows(vedtak.behandling.id)
-        val totrinnskontroll = totrinnskontrollService.finnAktivForBehandling(vedtak.behandling.id)
-        val enhetNavn = arbeidsfordelingService.hentArbeidsfordelingPåBehandling(vedtak.behandling.id).behandlendeEnhetNavn
-
-        return GrunnlagOgSignaturData(
-            grunnlag = personopplysningGrunnlag,
-            saksbehandler = totrinnskontroll?.saksbehandler ?: SikkerhetContext.hentSaksbehandlerNavn(),
-            beslutter = totrinnskontroll?.beslutter ?: "Beslutter",
-            enhet = enhetNavn
-        )
-    }
-
-    private fun hentHjemler(
-        behandlingId: Long,
-        utvidetVedtaksperioderMedBegrunnelser: List<UtvidetVedtaksperiodeMedBegrunnelser>,
-        målform: Målform,
-        sanityBegrunnelser: List<SanityBegrunnelse>
-    ): String {
-        val vilkårsvurdering = vilkårsvurderingService.hentAktivVilkårsvurderingForBehandling(behandlingId = behandlingId)
-
-        val opplysningspliktHjemlerSkalMedIBrev =
-            vilkårsvurdering.finnOpplysningspliktVilkår()?.resultat == Resultat.IKKE_OPPFYLT
-
-        return hentHjemmeltekst(
-            opplysningspliktHjemlerSkalMedIBrev = opplysningspliktHjemlerSkalMedIBrev,
-            målform = målform,
-            erFriteksterIPeriode = utvidetVedtaksperioderMedBegrunnelser.any { it.fritekster.isNotEmpty() },
-            sanitybegrunnelserBruktIBrev = utvidetVedtaksperioderMedBegrunnelser.flatMap { it.begrunnelser }
-                .mapNotNull { it.begrunnelse.tilSanityBegrunnelse(sanityBegrunnelser) }
-        )
-    }
-
-    private data class GrunnlagOgSignaturData(
-        val grunnlag: PersonopplysningGrunnlag,
-        val saksbehandler: String,
-        val beslutter: String,
-        val enhet: String
-    )
 }
