+++ src/test/enhetstester/kotlin/no/nav/familie/ks/sak/kjerne/brev/BrevServiceTest.kt
@@ -11,7 +11,6 @@ import io.mockk.slot
 import io.mockk.verify
 import no.nav.familie.kontrakter.felles.dokarkiv.v2.Førsteside
 import no.nav.familie.ks.sak.api.dto.ManueltBrevDto
-import no.nav.familie.ks.sak.common.exception.Feil
 import no.nav.familie.ks.sak.data.lagBehandling
 import no.nav.familie.ks.sak.data.lagFagsak
 import no.nav.familie.ks.sak.data.lagPerson
@@ -22,28 +21,19 @@ import no.nav.familie.ks.sak.data.shouldNotBeNull
 import no.nav.familie.ks.sak.integrasjon.familieintegrasjon.IntegrasjonClient
 import no.nav.familie.ks.sak.integrasjon.journalføring.UtgåendeJournalføringService
 import no.nav.familie.ks.sak.integrasjon.journalføring.domene.JournalføringRepository
-import no.nav.familie.ks.sak.integrasjon.sanity.SanityService
 import no.nav.familie.ks.sak.kjerne.arbeidsfordeling.ArbeidsfordelingService
 import no.nav.familie.ks.sak.kjerne.arbeidsfordeling.domene.ArbeidsfordelingPåBehandling
 import no.nav.familie.ks.sak.kjerne.behandling.SettBehandlingPåVentService
 import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingKategori
 import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingRepository
 import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingÅrsak
-import no.nav.familie.ks.sak.kjerne.behandling.steg.simulering.SimuleringService
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.VedtakService
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene.VedtaksperiodeService
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.VilkårsvurderingService
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Resultat
-import no.nav.familie.ks.sak.kjerne.brev.domene.maler.BrevDto
 import no.nav.familie.ks.sak.kjerne.brev.domene.maler.Brevmal
-import no.nav.familie.ks.sak.kjerne.brev.domene.maler.InnhenteOpplysningerDataDto
 import no.nav.familie.ks.sak.kjerne.logg.LoggService
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.PersonopplysningGrunnlagService
-import no.nav.familie.ks.sak.kjerne.totrinnskontroll.TotrinnskontrollService
 import no.nav.familie.prosessering.internal.TaskService
-import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.api.extension.ExtendWith
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
@@ -51,7 +41,7 @@ import org.junit.jupiter.params.provider.EnumSource
 @ExtendWith(MockKExtension::class)
 class BrevServiceTest {
     @MockK
-    private lateinit var brevKlient: BrevKlient
+    private lateinit var genererBrevService: GenererBrevService
 
     @MockK
     private lateinit var personopplysningGrunnlagService: PersonopplysningGrunnlagService
@@ -83,24 +73,6 @@ class BrevServiceTest {
     @MockK
     private lateinit var settBehandlingPåVentService: SettBehandlingPåVentService
 
-    @MockK
-    private lateinit var totrinnskontrollService: TotrinnskontrollService
-
-    @MockK
-    private lateinit var brevPeriodeService: BrevPeriodeService
-
-    @MockK
-    private lateinit var sanityService: SanityService
-
-    @MockK
-    private lateinit var vedtaksperiodeService: VedtaksperiodeService
-
-    @MockK
-    private lateinit var vedtakService: VedtakService
-
-    @MockK
-    private lateinit var simuleringService: SimuleringService
-
     @InjectMockKs
     private lateinit var brevService: BrevService
 
@@ -115,9 +87,7 @@ class BrevServiceTest {
         )
 
     @Test
-    fun `hentForhåndsvisningAvBrev - skal hente pdf i form av en ByteArray fra BrevKlient`() {
-        val brevSlot = slot<BrevDto>()
-
+    fun `hentForhåndsvisningAvBrev - skal hente pdf i form av en ByteArray fra genererBrevService`() {
         every { personopplysningGrunnlagService.hentSøker(behandlingId = behandling.id) } returns lagPerson(
             lagPersonopplysningGrunnlag(behandlingId = behandling.id, søkerPersonIdent = søker.aktivFødselsnummer()),
             søker
@@ -128,53 +98,9 @@ class BrevServiceTest {
             behandlendeEnhetId = "1234"
         )
 
-        every { brevKlient.genererBrev(any(), capture(brevSlot)) } returns ByteArray(10)
+        every { genererBrevService.genererManueltBrev(any(), any()) } returns ByteArray(10)
 
         brevService.hentForhåndsvisningAvBrev(behandlingId = behandling.id, manueltBrevDto).shouldNotBeNull()
-
-        val brev = brevSlot.captured
-
-        assertEquals(Brevmal.INNHENTE_OPPLYSNINGER, brev.mal)
-
-        val innhenteOpplysningerData = brev.data as InnhenteOpplysningerDataDto
-
-        assertEquals("Behandlende enhet", innhenteOpplysningerData.delmalData.signatur.enhet?.first())
-        assertEquals(søker.aktivFødselsnummer(), innhenteOpplysningerData.flettefelter.fodselsnummer?.first())
-        assertEquals(
-            "Dokumentasjon som viser når barna kom til Norge.",
-            innhenteOpplysningerData.flettefelter.dokumentliste?.first()
-        )
-    }
-
-    @Test
-    fun `hentForhåndsvisningAvBrev - skal kaste feil dersom kall mot 'familie-brev' feiler`() {
-        every { personopplysningGrunnlagService.hentSøker(behandlingId = behandling.id) } returns lagPerson(
-            lagPersonopplysningGrunnlag(behandlingId = behandling.id, søkerPersonIdent = søker.aktivFødselsnummer()),
-            søker
-        )
-        every { arbeidsfordelingService.hentArbeidsfordelingPåBehandling(behandlingId = behandling.id) } returns ArbeidsfordelingPåBehandling(
-            behandlingId = behandling.id,
-            behandlendeEnhetNavn = "Behandlende enhet",
-            behandlendeEnhetId = "1234"
-        )
-
-        every {
-            brevKlient.genererBrev(
-                any(),
-                any()
-            )
-        } throws Exception("Kall mot familie-brev feilet")
-
-        val feil =
-            assertThrows<Feil> { brevService.hentForhåndsvisningAvBrev(behandlingId = behandling.id, manueltBrevDto) }
-        assertEquals(
-            "Klarte ikke generere brev for ${manueltBrevDto.brevmal}. Kall mot familie-brev feilet",
-            feil.message
-        )
-        assertEquals(
-            "Det har skjedd en feil. Prøv igjen, og ta kontakt med brukerstøtte hvis problemet vedvarer.",
-            feil.frontendFeilmelding
-        )
     }
 
     @ParameterizedTest
@@ -198,7 +124,7 @@ class BrevServiceTest {
 
         every { behandlingRepository.hentBehandling(behandlingId = behandling.id) } returns behandling
 
-        every { brevKlient.genererBrev(any(), any()) } returns ByteArray(10)
+        every { genererBrevService.genererManueltBrev(any(), any()) } returns ByteArray(10)
 
         val førstesideSlot = slot<Førsteside>()
 
@@ -257,7 +183,7 @@ class BrevServiceTest {
 
         every { behandlingRepository.hentBehandling(behandlingId = behandling.id) } returns behandling
 
-        every { brevKlient.genererBrev(any(), any()) } returns ByteArray(10)
+        every { genererBrevService.genererManueltBrev(any(), any()) } returns ByteArray(10)
 
         every {
             utgåendeJournalføringService.journalførDokument(
@@ -296,39 +222,6 @@ class BrevServiceTest {
     @ParameterizedTest
     @EnumSource(
         value = Brevmal::class,
-        names = ["VEDTAK_FØRSTEGANGSVEDTAK", "VEDTAK_ENDRING", "VEDTAK_OPPHØRT", "VEDTAK_OPPHØR_MED_ENDRING", "VEDTAK_AVSLAG", "VEDTAK_FORTSATT_INNVILGET", "VEDTAK_KORREKSJON_VEDTAKSBREV", "VEDTAK_OPPHØR_DØDSFALL", "AUTOVEDTAK_BARN_6_OG_18_ÅR_OG_SMÅBARNSTILLEGG", "AUTOVEDTAK_NYFØDT_FØRSTE_BARN", "AUTOVEDTAK_NYFØDT_BARN_FRA_FØR"],
-        mode = EnumSource.Mode.INCLUDE
-    )
-    fun `genererOgSendBrev - skal ikke journalføre brev for brevmaler som ikke kan sendes manuelt`(brevmal: Brevmal) {
-        every { personopplysningGrunnlagService.hentSøker(behandlingId = behandling.id) } returns lagPerson(
-            lagPersonopplysningGrunnlag(behandlingId = behandling.id, søkerPersonIdent = søker.aktivFødselsnummer()),
-            søker
-        )
-        every { arbeidsfordelingService.hentArbeidsfordelingPåBehandling(behandlingId = behandling.id) } returns ArbeidsfordelingPåBehandling(
-            behandlingId = behandling.id,
-            behandlendeEnhetNavn = "Behandlende enhet",
-            behandlendeEnhetId = "1234"
-        )
-
-        every { behandlingRepository.hentBehandling(behandlingId = behandling.id) } returns behandling
-
-        every { brevKlient.genererBrev(any(), any()) } returns ByteArray(10)
-
-        val feil = assertThrows<Feil> {
-            brevService.genererOgSendBrev(
-                behandling.id,
-                ManueltBrevDto(
-                    brevmal = brevmal,
-                    mottakerIdent = søker.aktivFødselsnummer()
-                )
-            )
-        }
-        assertEquals("Kan ikke mappe fra manuel brevrequest til $brevmal.", feil.message)
-    }
-
-    @ParameterizedTest
-    @EnumSource(
-        value = Brevmal::class,
         names = ["INNHENTE_OPPLYSNINGER", "VARSEL_OM_REVURDERING"],
         mode = EnumSource.Mode.INCLUDE
     )
@@ -349,7 +242,7 @@ class BrevServiceTest {
 
         every { behandlingRepository.hentBehandling(behandlingId = behandling.id) } returns behandling
 
-        every { brevKlient.genererBrev(any(), any()) } returns ByteArray(10)
+        every { genererBrevService.genererManueltBrev(any(), any()) } returns ByteArray(10)
 
         every {
             utgåendeJournalføringService.journalførDokument(
@@ -416,7 +309,7 @@ class BrevServiceTest {
 
         every { behandlingRepository.hentBehandling(behandlingId = behandling.id) } returns behandling
 
-        every { brevKlient.genererBrev(any(), any()) } returns ByteArray(10)
+        every { genererBrevService.genererManueltBrev(any(), any()) } returns ByteArray(10)
 
         every {
             utgåendeJournalføringService.journalførDokument(
